chore(cleanup): drop dead reconcile path + inert materializeProfile (deadcode sweep)
Post config-overlay migration these had no production callers:
- ProfileManager.materializeProfile (+ its only helper resolvedGeoDataUrls); the apply path
  runs through ProfileOverlay.refreshFromStore instead.
- RuleApplyService.reconcileWithStoredState + syncStoredWithParsed + reclassifyBySubscription —
  the last remnants of the rule reconcile-merge the overlay replaced (SubscriptionUpdateMerge was
  already removed).

normalizeRuleOrder stays live; its two tests move to the new RuleOrderTest. RuleReconcileTest
(otherwise only exercised the dead helpers) is removed. Build + service unit tests green; -363 lines.
