+++ service/src/main/java/com/github/kr328/clash/service/util/RuleApplyService.kt
@@ -185,127 +185,11 @@ class RuleApplyService(
         return RuleDryRun(currentYaml, mergedYaml, normalized)
     }
 
-    /**
-     * Reconcile after a subscription refresh. Reads MANUAL rules from the saved state file
-     * (anything the user added through the UI), reads PROVIDER rules from the freshly fetched
-     * config, and re-applies them through the same merge code path used by applyState. It relies
-     * on structured MANUAL-vs-PROVIDER classification from the saved state file rather than
-     * heuristic line matching.
-     *
-     * Files are passed in explicitly because the caller runs against `processingDir`, not the
-     * usual `importedDir/<uuid>`.
-     *
-     * @return true if reconciliation rewrote the config; false on any failure (caller should
-     *         leave the freshly fetched config alone in that case).
-     */
-    fun reconcileWithStoredState(
-        configFile: File,
-        stateFile: File,
-        fetchedSubRuleKeys: Set<String> = emptySet(),
-    ): Boolean {
-        if (!configFile.isFile) return false
-        if (!stateFile.isFile) return false
-        return runCatching {
-            val configText = configFile.readText()
-            val snapshot = Clash.parseProfileSnapshot(configFile.parentFile!!)
-            val parsed = RuleMapper.parseStateFromSnapshot(snapshot)
-            val classifiedParsed = parsed.copy(
-                rules = reclassifyBySubscription(parsed.rules, fetchedSubRuleKeys),
-            )
-            val stored = stateJsonForReconcile.decodeFromString(
-                RuleState.serializer(),
-                stateFile.readText(),
-            )
-            val merged = syncStoredWithParsed(stored, classifiedParsed)
-            val normalized = merged.copy(rules = normalizeRuleOrder(merged.rules))
-            val geoDataUrls = GeoDataSources.resolve(
-                preset = serviceStore.geoDataSourcePreset,
-                customGeoIp = serviceStore.geoDataCustomGeoIp,
-                customGeoSite = serviceStore.geoDataCustomGeoSite,
-                customMmdb = serviceStore.geoDataCustomMmdb,
-                customAsn = serviceStore.geoDataCustomAsn,
-            )
-            val mergedYaml = RuleMapper.mergeStateIntoConfig(configText, normalized, geoDataUrls)
-            // Soft engine check (see dryRunState comment for the rationale).
-            Clash.validateProfileBytes(mergedYaml)?.let { engineError ->
-                Log.w("Engine flagged reconciled config (continuing anyway): $engineError")
-            }
-            configFile.writeText(mergedYaml)
-            stateFile.writeText(stateJsonForReconcile.encodeToString(RuleState.serializer(), normalized))
-            Log.d("Reconcile after subscription update applied ${normalized.rules.count { !it.deleted && it.enabled }} active rule(s)")
-            true
-        }.getOrElse {
-            Log.w("Reconcile after subscription update failed; keeping mergeAfterFetch output", it)
-            false
-        }
-    }
-
     private fun configFile(uuid: UUID): File? {
         val file = File(context.importedDir, "$uuid/config.yaml")
         return file.takeIf { it.isFile }
     }
 
-}
-
-/**
- * Reconcile the user's saved [stored] rule state with the [incoming] rules parsed
- * from a freshly fetched subscription. Identity is `(type,value,policy)` — NOT
- * order — so the user's enable/disable/delete decision on a PROVIDER rule re-binds
- * to the same rule after a refresh even if its position changed. MANUAL rules and
- * any disabled/deleted stored rule that vanished from the fetch are retained.
- *
- * Pure (no Context/engine) so it is unit-testable; `reconcileWithStoredState`
- * delegates here. Mirrors `RuleRepository.syncProviderRules` classification.
- */
-internal fun syncStoredWithParsed(stored: RuleState, incoming: RuleState): RuleState {
-    fun key(r: RuleItem) =
-        "${r.type.uppercase()},${r.value.uppercase()},${r.policy.uppercase()}"
-
-    // A deleted MANUAL rule is gone for good (no upstream sub to restore from), so it must
-    // not survive a subscription refresh either. Drop these before matching/retention so they
-    // can't resurface or stamp deleted=true onto a later same-key re-add. Mirrors
-    // RuleRepository.syncProviderRules.
-    val storedRules = stored.rules.filterNot { it.deleted && it.source == RuleSource.MANUAL }
-    val byKey = storedRules.associateBy(::key)
-    val mergedRules = incoming.rules.mapIndexed { index, rule ->
-        val old = byKey[key(rule)]
-        if (old != null) {
-            rule.copy(
-                id = old.id,
-                enabled = old.enabled,
-                deleted = old.deleted,
-                isRestorable = old.isRestorable || rule.isRestorable,
-                order = index,
-            )
-        } else {
-            rule.copy(order = index)
-        }
-    }.toMutableList()
-
-    val incomingKeys = incoming.rules.map(::key).toSet()
-    val retained = storedRules.filter { rule ->
-        key(rule) !in incomingKeys &&
-            (rule.deleted || !rule.enabled || rule.source == RuleSource.MANUAL)
-    }
-    retained.forEach { mergedRules += it.copy(order = mergedRules.size) }
-    return stored.copy(providers = incoming.providers, rules = mergedRules)
-}
-
-/**
- * Reclassify rules by ORIGIN using the `(type,value,policy)` keys of the freshly fetched
- * subscription: a rule whose key is in the fetch is PROVIDER (heals a wrongly-stored MANUAL
- * back to subscription); one absent from the fetch is a genuine user MANUAL rule the merge
- * preserved. Empty keys → leave as-is (caller had no fetch to compare). Pure → unit-testable.
- */
-internal fun reclassifyBySubscription(
-    rules: List<RuleItem>,
-    fetchedSubRuleKeys: Set<String>,
-): List<RuleItem> {
-    if (fetchedSubRuleKeys.isEmpty()) return rules
-    return rules.map { r ->
-        val k = "${r.type.uppercase()},${r.value.uppercase()},${r.policy.uppercase()}"
-        r.copy(source = if (k in fetchedSubRuleKeys) RuleSource.PROVIDER else RuleSource.MANUAL)
-    }
 }
 
 /**
